+++ src/App.js
@@ -4,27 +4,25 @@ import "./reset.css";
 import AddressInput from "./components/createdForm/AddressInput";
 import { ThemeProvider } from "styled-components";
 import theme from "./theme";
-import SubForm from "./components/createdForm/SubForm";
+import CreatedForm from "./components/createdForm/CreatedForm";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import DataList from "./components/dataList/DataList";
 import FormList from "./components/formList/FormList";
 
 function App() {
   return (
-      <ThemeProvider theme={theme}>
+    <ThemeProvider theme={theme}>
       <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<FormList />} />
-        <Route path="/dataList" element={<DataList />} />
-        {/*<SubForm />*/}
-        {/*<AddressInput />*/}
-        {/* <Route path="/form" element={<Form />}></Route>
-        {/* <Route path="/form/:id" element={<Form />}></Route>
-        <Route path="/createdForm" element={<CreatedForm />}></Route> */}
-      </Routes>
-    </BrowserRouter>
-      </ThemeProvider>
-
+        <Routes>
+          <Route path="/" element={<FormList />} />
+          <Route path="/dataList" element={<DataList />} />
+          {/*<AddressInput />*/}
+          {/* <Route path="/form" element={<Form />}></Route> */}
+          {/* <Route path="/form/:id" element={<Form />}></Route> */}
+          <Route path="/createdForm" element={<CreatedForm />}></Route>
+        </Routes>
+      </BrowserRouter>
+    </ThemeProvider>
   );
 }
